# refactor(scripts): improve Ollama performance evaluation