This PR introduces the following modifications:
- add HTTP evaluation to avoid Ollama stat reporting issues
- renamed script and benchmark result
- improved script output
```shell
pxlksr/opencodeinterpreter-ds:6.7b-Q4_K_M (5 runs) at 2024-03-27 18:35:42.451474+00:00
HTTP: ttft 180.28ms (± 54.44) | ingestion 76.27 tok/s (± 9.39) | generation 38.00 tok/s (± 1.43)
Ollama-reported: load duration 0.25ms (± 0.35) | ingestion 77.58 tok/s (± 9.46) | generation 38.33 tok/s (± 1.49)
```